Crate honeycomb_kernels

honeycomb-kernels

This crate implements usual meshing algorithms using combinatorial maps as the underlying mesh representation structure. The implementation of those are provided in the core crate of the honeycomb project.

Kernels

grisubal – Grid Submersion Algorithm

Re-exports

pub use grisubal::grisubal;

pub use grisubal::GrisubalError;

Modules

grisubal

GRISUBAL algorithm description & implementation

Enums

Clip

Post-processing clip operation.