honeycomb_kernels::remeshing

Enum EdgeCollapseError

pub enum EdgeCollapseError {
    FailedCoreOp(SewError),
    FailedDartRelease(DartReleaseError),
    NonCollapsibleEdge(&'static str),
    InvertedOrientation,
    NullEdge,
    BadTopology,
}

Error-modeling enum for edge collapse routine.

Variants

FailedCoreOp(SewError)

A core operation failed.

FailedDartRelease(DartReleaseError)

A discarded dart couldnโt be released.

NonCollapsibleEdge(&'static str)

The edge passed as argument cannot be collapsed due to constraints on its vertices.

InvertedOrientation

The structure after collapse contains a triangle with inverted orientation, making the geometry invalid.

NullEdge

The edge passed as argument is null.

BadTopology

One or both of the cells adjacent to the edge are not triangles.