honeycomb_kernels::remeshing

Function capture_geometry

pub fn capture_geometry<T: CoordsFloat>(
    file_path: impl AsRef<Path>,
    grid_cell_sizes: [T; 2],
    clip: Clip,
) -> Result<CMap2<T>, GrisubalError>

Capture the geometry specified as input using the grisubal algorithm.

This function follows roughly the same step as the grisubal algorithm, but differs in two instances:

The overlapping grid avoids landing Points of Interest on all edges of the grid (instead of only corners), meaning we do not remove any from the original geometry.

Points of Interest are used to associate VertexAnchor::Node values to some vertices of the mesh.

A complete classification of the resulting mesh can be obtained using the classify_capture function.

Return / Errors

This function returns a Result taking the following values:

Ok(CMap2) – Algorithm ran successfully.

Err(GrisubalError) – Algorithm encountered an issue. See GrisubalError for all possible errors.

Panics

This function may panic if the specified VTK file cannot be opened.