honeycomb_kernels

Module triangulation

Polygon triangulation functions

This module contains implementations of simple polygon triangulation methods. These are not meshing functions; our goal with these is to cut existing cells of an irregular mesh into triangular cells.

With consideration to the above, we implement two polygon triangulation methods:

fanning – two versions of this are implemented:

a defensive one where the function actively search for a valid vertex to fan from

a specific one which assume the cell is convex; it fans the polygon from its first vertex

ear clipping – this method isn’t algorithmically efficient, but (a) we operate on small cells, and (b) it covers our needs (non-fannable polygons without holes)

Enums

TriangulateError

Error-modeling enum for triangulation routines.

Functions

check_requirements

Checks if a face meets the requirements for triangulation.

earclip_cell_countercw

Triangulates a face using the ear clipping method.

earclip_cell_cw

Triangulates a face using the ear clipping method.

fan_cell

Triangulates a face using a fan triangulation method.

fan_convex_cell

Triangulates a face using a fan triangulation method.